refactor(cli): unify codegen CLIs by sharing questions and types from graphql-codegen
