+++ graphql/codegen/src/cli/index.ts
@@ -5,13 +5,13 @@
  * This is a thin wrapper around the core generate() function.
  * All business logic is in the core modules.
  */
-import { CLI, CLIOptions, Inquirerer, Question, getPackageJson } from 'inquirerer';
+import { CLI, CLIOptions, Inquirerer, getPackageJson } from 'inquirerer';
 
 import { generate } from '../core/generate';
 import { findConfigFile, loadConfigFile } from '../core/config';
 import { isMultiConfig } from '../types/config';
-import type { GenerateResult } from '../core/generate';
 import type { GraphQLSDKConfigTarget } from '../types/config';
+import { codegenQuestions, printResult, type CodegenAnswers } from './shared';
 
 const usage = `
 graphql-codegen - GraphQL SDK generator for Constructive databases
@@ -23,9 +23,6 @@ Source Options (choose one):
   -c, --config <path>           Path to config file
   -e, --endpoint <url>          GraphQL endpoint URL
   -s, --schema-file <path>      Path to GraphQL schema file
-  --pgpm-module-path <path>     Path to PGPM module directory
-  --pgpm-workspace-path <path>  Path to PGPM workspace (requires --pgpm-module-name)
-  --pgpm-module-name <name>     PGPM module name in workspace
 
 Database Options:
   --schemas <list>              Comma-separated PostgreSQL schemas
@@ -37,58 +34,12 @@ Generator Options:
   -o, --output <dir>            Output directory
   -t, --target <name>           Target name (for multi-target configs)
   -a, --authorization <token>   Authorization header value
-  --keep-db                     Keep ephemeral database after generation
   --dry-run                     Preview without writing files
   -v, --verbose                 Show detailed output
 
   -h, --help                    Show this help message
   --version                     Show version number
 `;
-
-const questions: Question[] = [
-  {
-    name: 'endpoint',
-    message: 'GraphQL endpoint URL',
-    type: 'text',
-    required: false,
-  },
-  {
-    name: 'output',
-    message: 'Output directory',
-    type: 'text',
-    required: false,
-    default: './generated',
-    useDefault: true,
-  },
-  {
-    name: 'reactQuery',
-    message: 'Generate React Query hooks?',
-    type: 'confirm',
-    required: false,
-    default: false,
-    useDefault: true,
-  },
-  {
-    name: 'orm',
-    message: 'Generate ORM client?',
-    type: 'confirm',
-    required: false,
-    default: false,
-    useDefault: true,
-  },
-];
-
-function printResult(result: GenerateResult): void {
-  if (result.success) {
-    console.log('[ok]', result.message);
-    if (result.tables?.length) {
-      console.log('Tables:', result.tables.join(', '));
-    }
-  } else {
-    console.error('x', result.message);
-    result.errors?.forEach((e) => console.error('  -', e));
-  }
-}
 
 export const commands = async (
   argv: Record<string, unknown>,
@@ -148,45 +99,28 @@ export const commands = async (
     return argv;
   }
 
-  // No config file - use CLI args with prompts
-  const answers = await prompter.prompt({
-    ...argv,
-    output: argv.output || argv.o,
-    endpoint: argv.endpoint || argv.e,
-  }, questions) as Record<string, unknown>;
+  // No config file - prompt for options using shared questions
+  const answers = await prompter.prompt<CodegenAnswers>(argv as CodegenAnswers, codegenQuestions);
 
-  // Build db config if pgpm options provided
-  const pgpmModulePath = argv['pgpm-module-path'] as string | undefined;
-  const pgpmWorkspacePath = argv['pgpm-workspace-path'] as string | undefined;
-  const pgpmModuleName = argv['pgpm-module-name'] as string | undefined;
-  const schemasArg = argv.schemas as string | undefined;
-  const apiNamesArg = argv['api-names'] as string | undefined;
-
-  const db = (pgpmModulePath || pgpmWorkspacePath) ? {
-    pgpm: {
-      modulePath: pgpmModulePath,
-      workspacePath: pgpmWorkspacePath,
-      moduleName: pgpmModuleName,
-    },
-    schemas: schemasArg ? schemasArg.split(',').map((s) => s.trim()) : undefined,
-    apiNames: apiNamesArg ? apiNamesArg.split(',').map((s) => s.trim()) : undefined,
-    keepDb: !!(argv['keep-db']),
+  // Build db config if schemas or apiNames provided
+  const db = (answers.schemas || answers.apiNames) ? {
+    schemas: answers.schemas,
+    apiNames: answers.apiNames,
   } : undefined;
 
   const result = await generate({
-    endpoint: (answers.endpoint || argv.endpoint || argv.e) as string | undefined,
-    schemaFile: (argv['schema-file'] || argv.s) as string | undefined,
+    endpoint: answers.endpoint,
+    schemaFile: answers.schemaFile,
     db,
-    output: (answers.output || argv.output || argv.o) as string | undefined,
-    authorization: (argv.authorization || argv.a) as string | undefined,
-    reactQuery: !!(answers.reactQuery ?? argv['react-query']),
-    orm: !!(answers.orm ?? argv.orm),
-    dryRun: !!argv['dry-run'],
-    verbose: !!argv.verbose || !!argv.v,
+    output: answers.output,
+    authorization: answers.authorization,
+    reactQuery: answers.reactQuery,
+    orm: answers.orm,
+    dryRun: answers.dryRun,
+    verbose: answers.verbose,
   });
 
   printResult(result);
-  if (!result.success) process.exit(1);
   prompter.close();
   return argv;
 };
